requirements.txt
httpx>=0.23.3
load_dotenv>=0.1.0
ujson>=5.7.0
